✨ Feat: 트레이닝 패키지 구조 변경 및 api 수정에 따른 솔로 트레이닝 기능 수정
