--- src/main/java/com/solit/sync2sing/domain/training/TrainingService.java
@@ -1,50 +1,0 @@
-package com.solit.sync2sing.domain.training;
-
-import com.solit.sync2sing.domain.training.dto.*;
-import com.solit.sync2sing.global.type.TrainingMode;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public abstract class TrainingService {
-
-    private final TrainingMode trainingMode;
-
-    public TrainingService(TrainingMode trainingMode) {
-        this.trainingMode = trainingMode;
-    }
-
-//    CurriculumListResponse generateTrainingCurriculum(GenerateCurriculumRequest generateCurriculumRequest) {
-//
-//    }
-//
-//    SessionDTO getSession(UserDetails userDetails) {
-//
-//    }
-//
-//    SessionDTO createSession(UserDetails userDetails, CreateSessionRequest createSessionRequest) {
-//
-//    }
-//
-//    void deleteSession(UserDetails userDetails) {
-//
-//    }
-//
-//    SongListDTO getSongList(String type) {
-//
-//    }
-//
-//    SongListDTO.SongDTO getSong(Long songId, String type) {
-//
-//    }
-//
-//    CurrentTrainingListDTO getCurrentTrainingList(UserDetails userDetails) {
-//
-//    }
-//
-//    TrainingDTO setTrainingProgress(UserDetails userDetails, Long trainingId, Integer progress) {
-//
-//    }
-//
-//    VocalAnalysisReportDTO generateVocalAnalysisReport(String recordingFileUrl, GenerateVocalAnalysisReportRequest generateVocalAnalysisReportRequest) {
-//
-//    }
-}
