+++ src/main/java/com/solit/sync2sing/domain/training/base/dto/CreateSessionRequest.java
@@ -1,0 +1,4 @@
+package com.solit.sync2sing.domain.training.base.dto;
+
+public class CreateSessionRequest {
+}
