+++ src/main/java/com/solit/sync2sing/domain/training/base/dto/SongListDTO.java
@@ -1,4 +1,4 @@
-package com.solit.sync2sing.domain.training.dto;
+package com.solit.sync2sing.domain.training.base.dto;
 
 public class SongListDTO {
