+++ src/main/java/com/solit/sync2sing/domain/training/common/dto/VocalAnalysisReportDTO.java
@@ -1,0 +1,4 @@
+package com.solit.sync2sing.domain.training.common.dto;
+
+public class VocalAnalysisReportDTO {
+}
