--- src/main/java/com/solit/sync2sing/domain/training/dto/SessionDTO.java
@@ -1,4 +1,0 @@
-package com.solit.sync2sing.domain.training.dto;
-
-public class SessionDTO {
-}
