+++ src/main/java/com/solit/sync2sing/domain/training/solo/controller/SoloTrainingController.java
@@ -1,9 +1,6 @@
 package com.solit.sync2sing.domain.training.solo.controller;
 
-import com.solit.sync2sing.domain.training.dto.CreateSessionRequest;
-import com.solit.sync2sing.domain.training.dto.GenerateCurriculumRequest;
-import com.solit.sync2sing.domain.training.dto.GenerateVocalAnalysisReportRequest;
-import com.solit.sync2sing.domain.training.solo.dto.SetTrainingProgressRequest;
+import com.solit.sync2sing.domain.training.base.dto.CreateSessionRequest;
 import com.solit.sync2sing.domain.training.solo.service.SoloTrainingService;
 import com.solit.sync2sing.global.response.ResponseCode;
 import com.solit.sync2sing.global.response.ResponseDTO;
@@ -12,7 +9,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 
 @RestController
 @RequestMapping("/solo-training")
@@ -26,18 +22,7 @@ public class SoloTrainingController {
         this.soloTrainingService = soloTrainingService;
     }
 
-    @PostMapping("/curriculum")
-    public ResponseEntity<ResponseDTO> generateTrainingCurriculum(
-            @RequestBody GenerateCurriculumRequest generateCurriculumRequest
-    ) {
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .body(new ResponseDTO(
-                ResponseCode.CURRICULUM_CREATED,
-                soloTrainingService.generateTrainingCurriculum(generateCurriculumRequest)
-            ));
-    }
-
+    
     @GetMapping("/session")
     public ResponseEntity<ResponseDTO> getSession(
             @AuthenticationPrincipal UserDetails userDetails
@@ -118,45 +103,4 @@ public class SoloTrainingController {
             ));
     }
 
-    @GetMapping("/training")
-    public ResponseEntity<ResponseDTO> getCurrentTrainingList(
-            @AuthenticationPrincipal UserDetails userDetails
-    ) {
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(new ResponseDTO(
-                ResponseCode.SOLO_CURRENT_TRAINING_LIST_FETCHED,
-                soloTrainingService.getCurrentTrainingList(userDetails)
-            ));
-    }
-
-    @PutMapping("/training/{trainingId}")
-    public ResponseEntity<ResponseDTO> setTrainingProgress(
-            @AuthenticationPrincipal UserDetails userDetails,
-            @PathVariable Long trainingId,
-            @RequestBody SetTrainingProgressRequest setTrainingProgressRequest
-    ) {
-        return ResponseEntity
-            .status(HttpStatus.OK)
-            .body(new ResponseDTO(
-                ResponseCode.SOLO_TRAINING_PROGRESS_UPDATED,
-                soloTrainingService.setTrainingProgress(userDetails, trainingId, setTrainingProgressRequest.getProgress())
-            ));
-    }
-
-    @PostMapping("/vocal-analysis")
-    public ResponseEntity<ResponseDTO> generateVocalAnalysisReport(
-            @RequestPart(value = "vocal_file") MultipartFile vocal_file,
-            @RequestPart(value = "data") GenerateVocalAnalysisReportRequest generateVocalAnalysisReportRequest
-    ) {
-        // TODO: 보컬 파일 S3 업로드
-        String recordingFileUrl = "";
-
-        return ResponseEntity
-            .status(HttpStatus.CREATED)
-            .body(new ResponseDTO(
-                ResponseCode.VOCAL_ANALYSIS_REPORT_CREATED,
-                soloTrainingService.generateVocalAnalysisReport(recordingFileUrl, generateVocalAnalysisReportRequest)
-            ));
-    }
 }
