+++ src/main/java/com/solit/sync2sing/global/response/ResponseCode.java
@@ -18,7 +18,11 @@ import org.springframework.http.HttpStatus;
 public enum ResponseCode {
 
     // 응답 코드 정의
+    CURRENT_TRAINING_LIST_FETCHED(HttpStatus.OK, "현재 진행 중인 트레이닝 목록 조회에 성공했습니다."),
     CURRICULUM_CREATED(HttpStatus.CREATED, "커리큘럼 추천 생성에 성공했습니다."),
+    TRAINING_PROGRESS_UPDATED(HttpStatus.OK, "트레이닝 진행 상황 업데이트에 성공했습니다."),
+    VOCAL_ANALYSIS_REPORT_CREATED(HttpStatus.CREATED, "보컬 분석 리포트 생성에 성공했습니다."),
+
     SOLO_TRAINING_SESSION_FETCHED(HttpStatus.OK, "솔로 트레이닝 세션 정보 조회에 성공했습니다."),
     SOLO_TRAINING_SESSION_CREATED(HttpStatus.CREATED, "솔로 트레이닝 세션 생성에 성공했습니다."),
     SOLO_TRAINING_SESSION_DELETED(HttpStatus.OK, "솔로 트레이닝 종료 및 관련 데이터(사용자 보컬 오디오 파일 및 진행 중인 훈련 데이터)가 삭제되었습니다."),
@@ -26,9 +30,7 @@ public enum ResponseCode {
     SOLO_ORIGINAL_SONG_LIST_FETCHED(HttpStatus.OK, "솔로 트레이닝 원곡 목록 조회에 성공했습니다."),
     SOLO_MR_SONG_FETCHED(HttpStatus.OK, "솔로 트레이닝 MR 곡 조회에 성공했습니다."),
     SOLO_ORIGINAL_SONG_FETCHED(HttpStatus.OK, "솔로 트레이닝 원곡 조회에 성공했습니다."),
-    SOLO_CURRENT_TRAINING_LIST_FETCHED(HttpStatus.OK, "현재 진행 중인 솔로 트레이닝 훈련 목록 조회에 성공했습니다."),
-    SOLO_TRAINING_PROGRESS_UPDATED(HttpStatus.OK, "솔로 트레이닝 진행 상황 업데이트에 성공했습니다."),
-    VOCAL_ANALYSIS_REPORT_CREATED(HttpStatus.CREATED, "보컬 분석 리포트 생성에 성공했습니다."),
+    
     DUET_AUDIO_MERGED(HttpStatus.CREATED, "듀엣 음원 병합에 성공했습니다."),
     PARTNER_APPLICATION_CREATED(HttpStatus.CREATED, "파트너 신청이 완료되었습니다."),
     RECEIVED_PARTNER_APPLICATIONS_FETCHED(HttpStatus.OK, "받은 파트너 신청 목록 조회에 성공했습니다."),
